Switch to ESP-IDF framework with Arduino component (v0.02)
Build ESP-IDF from source via `framework = espidf, arduino` (pinned
espressif32@6.12.0 / IDF 4.4.7) for full SDK configuration control.
Custom app_main() replaces CONFIG_AUTOSTART_ARDUINO. LittleFS provided
via esp_littlefs component (v1.14.8 submodule). sdkconfig.defaults
enables LWIP_STATS and sets FreeRTOS tick rate.

Enable MIB2_STATS for per-interface TX/RX byte counters in wifi command.
Rename mqtt_client.h/cpp to conez_mqtt.h/cpp to avoid ESP-IDF collision.
Fix format-truncation, misleading-indentation, and class-memaccess
warnings from stricter ESP-IDF compiler flags. Bump version to v0.02.

+++ firmware/src/http/fwupdate.cpp
@@ -106,13 +106,13 @@ void dump_nvs()
       case NVS_TYPE_I32: {
         int32_t val;
         nvs_get_i32(handle, info.key, &val);
-        Serial.printf("int32_t   = %ld\n", val);
+        Serial.printf("int32_t   = %d\n", (int)val);
         break;
       }
       case NVS_TYPE_U32: {
         uint32_t val;
         nvs_get_u32(handle, info.key, &val);
-        Serial.printf("uint32_t  = %lu\n", val);
+        Serial.printf("uint32_t  = %u\n", (unsigned)val);
         break;
       }
       case NVS_TYPE_I64: {
